+++ src/datatypes/wrapper/sum.rs
@@ -84,7 +84,6 @@
 
 use crate::traits::functor::Functor;
 use crate::traits::hkt::HKT;
-use crate::traits::identity::Identity;
 use crate::traits::monoid::Monoid;
 use crate::traits::semigroup::Semigroup;
 use std::fmt;
@@ -437,16 +436,6 @@ impl<T: Clone + Add<Output = T> + Default> Monoid for Sum<T> {
 impl<T> HKT for Sum<T> {
     type Source = T;
     type Output<U> = Sum<U>;
-}
-
-impl<T: Clone + Add<Output = T>> Identity for Sum<T> {
-    fn value(&self) -> &Self::Source {
-        &self.0
-    }
-
-    fn into_value(self) -> Self::Source {
-        self.0
-    }
 }
 
 impl<T: Clone + Add<Output = T>> Functor for Sum<T> {